style: tweaks
